Check for successful status codes before continuing request
Sometimes the requests cannot complete due to a number of factors such
as bad auth, wrong expected data, etc. and if that happens, the request
needs to terminate immediately instead of continuing on to fail.

This isn't a big deal just checking for a 200 at the moment as Acquia
do not utiltise status codes like 201 for newly created resources
however that will be something to keep an eye on in case it does happen.